fix(adapters): use representation-specific resource accounting
Count BIDS physiology as f64 storage and XDF channel bookkeeping by active
value representation. Preserve total-allocation rejection and pin XDF's
deduced first-sample timing in regression coverage.

AI-Assisted-By: {"agent":"root","id":"ai:openai-codex-root","model":"not-disclosed-by-runtime","product":"Codex","provider":"OpenAI","roles":["author","tester"]}
File-Contribution: {"by":[{"actor":"ai:openai-codex-root","role":"author"}],"operation":"modify","path":"crates/lamquant-standard-adapters/src/bids_full.rs","summary":"Count retained physiology samples using their actual f64 representation."}
File-Contribution: {"by":[{"actor":"ai:openai-codex-root","role":"author"}],"operation":"modify","path":"crates/lamquant-standard-adapters/src/xdf.rs","summary":"Measure channel bookkeeping using the active typed value-vector representation."}
File-Contribution: {"by":[{"actor":"ai:openai-codex-root","role":"tester"}],"operation":"modify","path":"crates/lamquant-standard-adapters/tests/xdf_adapter.rs","summary":"Adjusted total bookkeeping bomb threshold and pinned deduced first-sample timing."}

+++ crates/lamquant-standard-adapters/src/bids_full.rs
@@ -690,7 +690,7 @@ fn parse_bids(
             retained_signal_bytes = retained_signal_bytes
                 .checked_add(
                     values
-                        .checked_mul(core::mem::size_of::<i64>() as u64)
+                        .checked_mul(core::mem::size_of::<f64>() as u64)
                         .ok_or(AdapterError::SourceTooLarge)?,
                 )
                 .ok_or(AdapterError::SourceTooLarge)?;
